requirements.txt
streamlit
plotly
matplotlib
numpy
anthropic
python-dotenv
requests